LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-self-baseline-not-applied-if-sizing-cyclic-dependency-003-expected.txt

flex row
É
É
É
 flex row
É
É
É
 flex row
É
É
É
 flex column - all items share a baseline group
É
É
É
 flex column - all items share a baseline group
É
É
É
 flex column - all items share a baseline group
É
É
É

PASS .grid 1
PASS .grid 2
PASS .grid 3
PASS .grid 4
FAIL .grid 5 assert_equals:
<div class="grid column justifyItemsBaseline">
  <div class="item1 relativeWidth height50" data-offset-x="0" data-offset-y="0" data-expected-width="40" data-expected-height="50">É</div>
  <div class="item2 verticalLR" data-offset-x="12" data-offset-y="50" data-expected-width="20" data-expected-height="50">É</div>
  <div class="item3 verticalLR" data-offset-x="0" data-offset-y="100" data-expected-width="80" data-expected-height="100">É</div>
</div>
offsetLeft expected 0 but got 16
FAIL .grid 6 assert_equals:
<div class="grid column justifyItemsBaseline">
  <div class="item1 relativeWidth height50" data-offset-x="0" data-offset-y="0" data-expected-width="40" data-expected-height="50">É</div>
  <div class="item2" data-offset-x="16" data-offset-y="50" data-expected-width="20" data-expected-height="50">É</div>
  <div class="item3 verticalLR" data-offset-x="0" data-offset-y="100" data-expected-width="80" data-expected-height="100">É</div>
</div>
offsetLeft expected 0 but got 16
